cross domain dev fix, implemented register roundtrip
