+++ @app/client/package.json
@@ -14,6 +14,7 @@
     "apollo-link-http": "^1.5.16",
     "apollo-link-ws": "^1.0.19",
     "graphql": "^14.5.8",
+    "graphql.macro": "^1.4.2",
     "polished": "^3.4.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
